Decoupled gui res from gpu instance
The GPUInstance struct was messy with lots of
optional fields for when a GUI is spawned (which
is optional) to display the models used in
generating the tool paths, so I extracted them
and put them in a new struct specifically for them.
That struct is now an optional return from my
function that creates the GPUInstance. This seems
to solve a lot of memory ownership issues as well,
since I only want a single thread to handle
those GUI resources, and not tag along with the
other basic GPU Vulkan instance information
and queues, since path generation and use
of supplimentary graphics and compute
pipelines will eventually be on a different
frame so the GUI is not blocked when
computing paths.